+++ modules/clean.py
@@ -37,10 +37,26 @@ def clean_website_build():
     resources_path = os.path.join(config.resources_markdown_path, "resources.md")
     if os.path.isfile(resources_path):
         os.remove(resources_path)
-           # Remove resources markdown file
+    # Remove attackcon markdown file
+    attackcon_path = os.path.join(config.resources_markdown_path, "attackcon.md")
+    if os.path.isfile(attackcon_path):
+        os.remove(attackcon_path)
+    # Remove previous markdown file
     previous_path = os.path.join(config.resources_markdown_path, "previous.md")
     if os.path.isfile(previous_path):
         os.remove(previous_path)
+    # remove FAQ markdown file
+    faq_path = os.path.join(config.resources_markdown_path, "faq.md")
+    if os.path.isfile(faq_path):
+        os.remove(faq_path)
+    # Remove training markdown file
+    training_path = os.path.join(config.resources_markdown_path, "training.md")
+    if os.path.isfile(training_path):
+        os.remove(training_path)
+    # Remove training markdown file
+    training_cti_path = os.path.join(config.resources_markdown_path, "training_cti.md")
+    if os.path.isfile(training_cti_path):
+        os.remove(training_cti_path)
 
     # Remove output directory
     output_path = "output"